# src/main/java/com/rozetka/dto/response/PurchaseResponseDTO.java
package com.rozetka.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class PurchaseResponseDTO {

    private String orderTrackingNumber;
}